fitness is now long rather than int
It's like I flipflop for fun

+++ src/natural/ACO/Colony.java
@@ -80,7 +80,7 @@ public class Colony extends AbstractPopulation {
     }
 
     @Override
-    public int getBestFitness() {
+    public long getBestFitness() {
         return best.getFitness();
     }
 }
