+++ src/natural/GA/preCalc/PreCalcInterface.java
@@ -3,5 +3,5 @@ package natural.GA.preCalc;
 import natural.GA.Individual;
 
 public interface PreCalcInterface {
-    public int[] calc(Individual[] population);
+    public long[] calc(Individual[] population);
 }
